pyproject.toml
[project]
name = "ai-intel-pipeline"
version = "0.1.0"
description = "AI intel pipeline (no-make): fetch, normalize, gate, store, deliver"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "PyYAML>=6.0.1",
  "python-dateutil>=2.9.0",
  "feedparser>=6.0.11",
  "youtube-transcript-api>=0.6.2",
  "tenacity>=8.2.3",
  "colorama>=0.4.6",
  "typer>=0.12.3",
  "rich>=13.8.0",
]

[project.scripts]
ai-intel = "ai_intel_pipeline.cli:app"

